# docs/swagger.yaml
basePath: /
definitions:
  dto.DecodeURLResp:
    properties:
      url:
        type: string
    type: object
  dto.EchoHttpErrorResp:
    properties:
      message: {}
    type: object
  dto.EncodeURLResp:
    properties:
      url:
        type: string
    type: object
  httpurlshortener.DecodeURLInput:
    properties:
      url:
        type: string
    required:
    - url
    type: object
  httpurlshortener.EncodeURLInput:
    properties:
      url:
        type: string
    required:
    - url
    type: object
host: localhost:8081
info:
  contact: {}
  description: API for encoding and decoding short URLs.
  title: URL Shortener API
  version: "1.0"
paths:
  /url-shortener/decode:
    post:
      consumes:
      - application/json
      description: Decode a given URL to a shortened version.
      parameters:
      - description: URL Input to be decoded
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/httpurlshortener.DecodeURLInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DecodeURLResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.EchoHttpErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.EchoHttpErrorResp'
      summary: Decode URL
      tags:
      - url-shortener
  /url-shortener/encode:
    post:
      consumes:
      - application/json
      description: Encode a given URL to a shortened version.
      parameters:
      - description: URL Input to be encoded
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/httpurlshortener.EncodeURLInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EncodeURLResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.EchoHttpErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.EchoHttpErrorResp'
      summary: Encode URL
      tags:
      - url-shortener
swagger: "2.0"